+++ CHANGELOG.md
@@ -6,9 +6,8 @@ Version 0.9alpha
 - support for foreign data wrappers
 - support for foreign servers
 - support for foreign user mappings
-- !support for text search configurations (regconfig)
-- !support for text search dictionaries (regdictionary)
-
+- support for text search configurations (regconfig)
+- support for text search dictionaries (regdictionary)
 
 Version 0.8
 -----------